component artifacts now in json manifest and new script for downloading artifacts
