+++ .gitignore
@@ -1,4 +1,5 @@
 *~
+product-base/component_info
 
 core/zenpacks
 core/Dockerfile
